pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[project]
name = "inep-bibliometric-tool"
version = "0.1.0"
description = "Coverage, quality, reliability and accessibility assessment of bibliometric and innovation sources for Brazilian HEIs"
requires-python = ">=3.11"
dependencies = [
    # HTTP & APIs
    "httpx>=0.27",
    "tenacity>=8.2",          # retry logic
    # Data
    "pandas>=2.2",
    "pyarrow>=16.0",          # parquet support
    "SQLAlchemy>=2.0",
    # Schema & validation
    "pydantic>=2.7",
    # Matching & fuzzy search
    "rapidfuzz>=3.9",
    # Statistics
    "statsmodels>=0.14",      # Wilson CI
    # Config
    "pyyaml>=6.0",
    # Scraping
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    # Reporting
    "jinja2>=3.1",
    "markdown>=3.6",
    # Dashboard
    "streamlit>=1.36",
    "plotly>=5.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-httpx>=0.30",     # mock HTTP in tests
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["integration: live API tests, require credentials and network access"]
